# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Document Management Service API with PostgreSQL, Redis, Cassandra, and S3 datastores.",
        "title": "Doc Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/documents": {
            "post": {
                "description": "Accepts a file (CSV, PDF, etc.) with metadata, stores record in DB, prevents duplicate uploads via content_hash, and starts async processing",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Upload and accept a document",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Document file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MIME type or file format (e.g. pdf, csv)",
                        "name": "typeOfFile",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Optional metadata in JSON or plain text format",
                        "name": "optionalMeta",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document accepted successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apihelpers.ApiRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Document"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing file or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.ApiRes"
                        }
                    },
                    "409": {
                        "description": "Duplicate document: identical name and content",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.ApiRes"
                        }
                    },
                    "500": {
                        "description": "Server or database error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.ApiRes"
                        }
                    }
                }
            }
        },
        "/documents/{id}/events": {
            "get": {
                "description": "Queries Cassandra for all historical event snapshots and status transitions for a given document UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get document event history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apihelpers.ApiRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.DocumentSnapshot"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid document ID format",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.ApiRes"
                        }
                    },
                    "500": {
                        "description": "Server or database error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.ApiRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apihelpers.ApiRes": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.Document": {
            "type": "object",
            "properties": {
                "content_hash": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "extracted_content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optional_meta": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DocumentSnapshot": {
            "type": "object",
            "properties": {
                "db_snapshot": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}